\docType{package}
\name{unifiedWMWqPCR-package}
\alias{uWMWqPCR}
\alias{unifiedWMWqPCR}
\alias{unifiedWMWqPCR-package}
\title{Unified Wilcoxon-Mann Whitney Test for qPCR data.}
\description{
This package implements the unified Wilxocon-Mann Whitney Test for 
qPCR data. For more information, check the vignette.

When used for published research, please refer to the following
Bio-informatics paper:

De Neve, Jan; Meys, Joris; Ottoy, Jean-Pierre; Clement, Lieven and Thas,
  Olivier. "UnifiedWMWqPCR: The unified Wilcoxon-Mann-Whitney test for
  analyzing RT-qPCR data in R." Bioinformatics(2014): Vol.30, Issue 17,
  pp.2494-2495.

See also: \code{citation("unifiedWMWqPCR")}
}

