\docType{package}
\name{unifiedWMWqPCR-package}
\alias{uWMWqPCR}
\alias{unifiedWMWqPCR}
\alias{unifiedWMWqPCR-package}
\title{Unified Wilcoxon-Mann Whitney Test for qPCR data.}
\description{
\tabular{ll}{ Package: \tab unifiedWMWqPCR\cr Type: \tab
Package\cr Version: \tab 0.99.4\cr Date: \tab 2014-02-21\cr
License: \tab GPL-2\cr }
}

