% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifiedWMWqPCR-package.R
\docType{package}
\name{unifiedWMWqPCR-package}
\alias{unifiedWMWqPCR-package}
\alias{unifiedWMWqPCR}
\alias{uWMWqPCR}
\title{Unified Wilcoxon-Mann Whitney Test for qPCR data.}
\description{
This package allows you to carry out the unified Wilcoxon-Mann Whitney
Test for qPCR data. If you use this package for research purposes,
please cite the following article:
}
\details{
De Neve, Jan; Meys, Joris; Ottoy, Jean-Pierre; Clement, Lieven and Thas,
Olivier. "UnifiedWMWqPCR: The unified Wilcoxon-Mann-Whitney test for
analyzing RT-qPCR data in R." Bioinformatics(2014): Vol.30, Issue 17,
pp.2494-2495.

Or see \code{citation("unifiedWMWqPCR")}
}
